chore: remove v2 references from document data access service and its usages
